requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
plotly
wikipedia
pyrebase4
